API protection (must be admin to use any API) working!!!

+++ src/pages/api/delete-machine.ts
@@ -1,11 +1,17 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth";
 import prisma from "services/prisma";
 import { prismaErrHandler } from "services/prismaErrHandler";
+import { authOptions } from "./auth/[...nextauth]";
 
 export default async function handle(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  //--API Protection--
+  const session = await getServerSession(req, res, authOptions);
+  if (!session?.user.isAdmin) return res.status(401).json("Unauthorized");
+  //--ID based deletion--
   const { id } = req.body;
   try {
     const op = await prisma.machine.delete({
